feat(tax-reports): add reports page with DATEV + CSV export
Phase 11 complete — Milestone v1.5 Steuer-Berichte done:
- Reports page with date range picker + presets
- Payment method filter toggles (Cash/Cashless/Card)
- VAT breakdown table (per tax rate: gross, net, tax, count)
- Summary KPIs (total gross, sales count, avg, tax, net)
- Simple CSV export (UTF-8 BOM, German locale, semicolons)
- DATEV EXTF Buchungsstapel export (Windows-1252, v700, SKR03)
- Tax readiness blocker with lazy data loading guard
- Sidebar navigation link
- Fix ::numeric cast in backfill + stamp trigger functions
- Updated seed.sql with realistic multi-machine test data

+++ Docker/supabase/migrations/20260406000000_tax_infrastructure.sql
@@ -235,10 +235,10 @@ BEGIN
 
     -- 3d. Calculate and stamp if rate found
     IF v_tax_rate IS NOT NULL THEN
-      v_price_net := ROUND(NEW.item_price / (1 + v_tax_rate), 4);
+      v_price_net := ROUND(NEW.item_price::numeric / (1 + v_tax_rate), 4);
       NEW.tax_rate_snapshot := v_tax_rate;
       NEW.price_net := v_price_net;
-      NEW.tax_amount := ROUND(NEW.item_price - v_price_net, 4);
+      NEW.tax_amount := ROUND(NEW.item_price::numeric - v_price_net, 4);
     END IF;
     -- If no rate found, all three columns remain NULL (graceful fallback)
   END IF;
